Deduplicate vault services, example adapters, and server mocks; set jscpd threshold to 0%
