+++ .changeset/code-health-dedup-mpc-types.md
@@ -1,0 +1,5 @@
+---
+"@vultisig/mpc-types": patch
+---
+
+Extract `MpcPerAlgorithmEngineBase` so DKLS and Schnorr engine typings stay aligned without duplicated blocks.
