+++ examples/shared/src/adapters/index.ts
@@ -1,2 +1,3 @@
 export * from './context'
+export { SDKAdapterEventSubscriptionBase } from './SDKAdapterEventSubscriptionBase'
 export * from './types'
